Переносим отрисовку карточки фильма и его попапа в 'FilmController'.
